pyproject.toml
[project]
name = "omnibooker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "cryptography>=45.0.7",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "pytz>=2025.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
]
